destroy: aws: move elb helper functions to their own file
